destructure router component
